Provided support for the Max Function and moved appropriate tests from MathFunctionTests into MaxTests.
