+++ EPPlus/FormulaParsing/Excel/Functions/Math/Max.cs
@@ -31,18 +31,28 @@
 using System.Collections.Generic;
 using System.Linq;
 using OfficeOpenXml.FormulaParsing.ExpressionGraph;
+using OfficeOpenXml.Utils;
 
 namespace OfficeOpenXml.FormulaParsing.Excel.Functions.Math
 {
+	/// <summary>
+	/// This class contains the formula for calculating the maximum item in an array, list, or excel range.
+	/// </summary>
 	public class Max : HiddenValuesHandlingFunction
 	{
+		/// <summary>
+		/// Takes the user specified arguments and returns the maximum value.
+		/// </summary>
+		/// <param name="arguments">The user specified array, list, or excel range to take the maximum of.</param>
+		/// <param name="context">The context in which the program is being run.</param>
+		/// <returns>The maximum item of the user specified arguments.</returns>
 		public override CompileResult Execute(IEnumerable<FunctionArgument> arguments, ParsingContext context)
 		{
 			if (ValidateArguments(arguments, 1) == false)
 				return new CompileResult(eErrorType.Value);
 			var args = arguments.ElementAt(0);
 			var argumentValueList = this.ArgsToObjectEnumerable(false, new List<FunctionArgument> { args }, context);
-			var values = argumentValueList.Where(arg => arg.GetType().IsPrimitive && (arg is bool == false));
+			var values = argumentValueList.Where(arg => ((arg.GetType().IsPrimitive && (arg is bool == false)) || arg is System.DateTime));
 			if (arguments.ElementAt(0).Type.Name.Equals("List`1"))
 			{
 				if (values.Count() > 255)
@@ -51,7 +61,24 @@ namespace OfficeOpenXml.FormulaParsing.Excel.Functions.Math
 			}
 			else if (!arguments.ElementAt(0).IsExcelRange)
 			{
-				var tvalues = ArgsToDoubleEnumerable(IgnoreHiddenValues, false, arguments, context);
+				var tvalues = new List<double> { };
+				foreach (var item in arguments)
+				{
+					if (item.ExcelStateFlagIsSet(ExcelCellState.HiddenCell))
+						continue;
+					if (item.Value is string)
+					{
+						if (ConvertUtil.TryParseNumericString(item.Value, out double result))
+							tvalues.Add(result);
+						else if (ConvertUtil.TryParseDateString(item.Value, out System.DateTime res))
+						{
+							var temp = res.ToOADate();
+							tvalues.Add(temp);
+						}
+					}
+					else
+						tvalues.Add(ArgToDecimal(item.Value));
+				}
 				if (tvalues.Count() == 0)
 					return new CompileResult(eErrorType.Value);
 				if (tvalues.Count() > 255)
